+++ xarray/core/common.py
@@ -576,7 +576,10 @@ class BaseDataObject(AttrAccessMixin):
             outobj = self
             outcond = cond
 
-        return outobj._where(outcond)
+        # preserve attributes
+        out = outobj._where(outcond)
+        out._copy_attrs_from(self)
+        return out
 
     def close(self):
         """Close any files linked to this object