+++ xarray/core/dataarray.py
@@ -1001,7 +1001,9 @@ class DataArray(AbstractArray, BaseDataObject):
         if utils.is_dict_like(value):
             raise TypeError('cannot provide fill value as a dictionary with '
                             'fillna on a DataArray')
-        return self._fillna(value)
+        out = self._fillna(value)
+        out.attrs = self.attrs
+        return out
 
     def reduce(self, func, dim=None, axis=None, keep_attrs=False, **kwargs):
         """Reduce this array by applying `func` along some dimension(s).
@@ -1414,6 +1416,9 @@ class DataArray(AbstractArray, BaseDataObject):
                 f(self.variable, other_variable)
             return self
         return func
+
+    def _copy_attrs_from(self, other):
+        self.attrs = other.attrs
 
     @property
     def plot(self):