+++ xarray/test/test_dataset.py
@@ -2263,6 +2263,25 @@ class TestDataset(TestCase):
         actual = ds.groupby('b').fillna(Dataset({'a': ('b', [0, 2])}))
         self.assertDatasetIdentical(expected, actual)
 
+        # attrs with groupby
+        ds.attrs['attr'] = 'ds'
+        ds.a.attrs['attr'] = 'da'
+        actual = ds.groupby('b').fillna(Dataset({'a': ('b', [0, 2])}))
+        self.assertEqual(actual.attrs, ds.attrs)
+        self.assertEqual(actual.a.name, 'a')
+        self.assertEqual(actual.a.attrs, ds.a.attrs)
+
+        da = DataArray(range(5), name='a', attrs={'attr':'da'})
+        actual = da.fillna(1)
+        self.assertEqual(actual.name, 'a')
+        self.assertEqual(actual.attrs, da.attrs)
+
+        ds = Dataset({'a': da}, attrs={'attr':'ds'})
+        actual = ds.fillna({'a': 1})
+        self.assertEqual(actual.attrs, ds.attrs)
+        self.assertEqual(actual.a.name, 'a')
+        self.assertEqual(actual.a.attrs, ds.a.attrs)
+
     def test_where(self):
         ds = Dataset({'a': ('x', range(5))})
         expected = Dataset({'a': ('x', [np.nan, np.nan, 2, 3, 4])})
@@ -2296,6 +2315,26 @@ class TestDataset(TestCase):
         expected['a'].values = [0, 1] + [np.nan] * 3
         actual = ds.groupby('c').where(cond)
         self.assertDatasetIdentical(expected, actual)
+
+        # attrs with groupby
+        ds.attrs['attr'] = 'ds'
+        ds.a.attrs['attr'] = 'da'
+        actual = ds.groupby('c').where(cond)
+        self.assertEqual(actual.attrs, ds.attrs)
+        self.assertEqual(actual.a.name, 'a')
+        self.assertEqual(actual.a.attrs, ds.a.attrs)
+
+        # attrs
+        da = DataArray(range(5), name='a', attrs={'attr':'da'})
+        actual = da.where(da.values > 1)
+        self.assertEqual(actual.name, 'a')
+        self.assertEqual(actual.attrs, da.attrs)
+
+        ds = Dataset({'a': da}, attrs={'attr':'ds'})
+        actual = ds.where(ds > 0)
+        self.assertEqual(actual.attrs, ds.attrs)
+        self.assertEqual(actual.a.name, 'a')
+        self.assertEqual(actual.a.attrs, ds.a.attrs)
 
     def test_where_drop(self):
         # if drop=True